Enable Armeria server instrumentation notFound() test
