+++ instrumentation/armeria-1.3/library/src/main/java/io/opentelemetry/instrumentation/armeria/v1_3/ArmeriaNetClientAttributesExtractor.java
@@ -3,27 +3,27 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.armeria.v1_3.internal;
+package io.opentelemetry.instrumentation.armeria.v1_3;
 
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.logging.RequestLog;
-import io.opentelemetry.instrumentation.api.instrumenter.net.InetSocketAddressNetServerAttributesExtractor;
+import io.opentelemetry.instrumentation.api.instrumenter.net.InetSocketAddressNetClientAttributesExtractor;
 import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
 import javax.annotation.Nullable;
 
-public final class ArmeriaNetServerAttributesExtractor
-    extends InetSocketAddressNetServerAttributesExtractor<RequestContext, RequestLog> {
+final class ArmeriaNetClientAttributesExtractor
+    extends InetSocketAddressNetClientAttributesExtractor<RequestContext, RequestLog> {
 
   @Override
-  public String transport(RequestContext ctx) {
+  public String transport(RequestContext ctx, @Nullable RequestLog requestLog) {
     return SemanticAttributes.NetTransportValues.IP_TCP;
   }
 
   @Override
   @Nullable
-  public InetSocketAddress getAddress(RequestContext ctx) {
+  public InetSocketAddress getAddress(RequestContext ctx, @Nullable RequestLog requestLog) {
     SocketAddress address = ctx.remoteAddress();
     if (address instanceof InetSocketAddress) {
       return (InetSocketAddress) address;
